No reply pings 4.

+++ robocop_ng/cogs/logs.py
@@ -227,10 +227,15 @@ class Logs(Cog):
         if message.reference is not None:
             if message.reference.resolved.author.id is not message.author.id:
                 if message.reference.resolved.author in message.mentions:
-                    for r in message.reference.resolved.author.roles:
-                        if r.id == 1059460475588448416:
-                            await message.reply(content=f"**{message.author.display_name}, do not reply ping users who don't want to be pinged.**\nPlease check if a user has a `No Reply Pings` role on them before pinging them in replies.\nYou can turn off reply pings by using the blue `@ ON` button to the right of the message bar.", mention_author=True)
-                            await message.delete()
+                    staffuser = False
+                    for r in message.author.roles:
+                        if r.id == 259199371361517569 or r.id == 256985367977263105:
+                            staffuser = True
+                    if staffuser = False:
+                        for r in message.reference.resolved.author.roles:
+                            if r.id == 1059460475588448416:
+                                await message.reply(content=f"**{message.author.display_name}, do not reply ping users who don't want to be pinged.**\nPlease check if a user has a `No Reply Pings` role on them before pinging them in replies.\nYou can turn off reply pings by using the blue `@ ON` button to the right of the message bar.", mention_author=True)
+                                await message.delete()
                         
         if message.channel.id not in config.spy_channels:
             return
